implement transaction for save blood stock details during the save blood test details

+++ src/main/java/lk/ijse/gdse/bbms/controller/BloodTestPageController.java
@@ -44,7 +44,7 @@ public class BloodTestPageController implements Initializable {
     private TableColumn<BloodTestTM, String> colSerialNum;
 
     @FXML
-    private TableColumn<BloodTestTM,Double> colPlatelets;
+    private TableColumn<BloodTestTM,Float> colPlatelets;
 
     @FXML
     private TableColumn<BloodTestTM,Double> colRedBloodCells;
